refactor(settings): split tab-app into audio + accessibility sections
